Prints size ofdifferent data types

+++ 0x00-hello_world/6-size.c
@@ -14,11 +14,11 @@ int main(void)
 	long int w;
 	double u;
 	
-	printf("Size of a char: %lu byte(s)\n", (unsigned long) sizeof(y));
-	printf("Size of a int: %lu byte(s)\n", (unsigned long) sizeof(x));
-	printf("Size of a long int: %lu byte(s)\n", (unsigned long) sizeof(w));
-	printf("Size of a long long int: %lu byte(s)\n", (unsigned long) sizeof(long w));
-	printf("Size of a float: %lu byte(s)\n", (unsigned long) sizeof((float) u));
+	printf("Size of a char: %d byte(s)\n", sizeof(y));
+	printf("Size of a int: %d byte(s)\n", sizeof(x));
+	printf("Size of a long int: %d byte(s)\n", sizeof(w));
+	printf("Size of a long long int: %d byte(s)\n", sizeof(long w));
+	printf("Size of a float: %d byte(s)\n", sizeof((float) u));
 
 	return (0);
 }
